Capture SIGSEGV signals
Add SIGSEGV singal handler to capture signal and print a visible mesage letting
the user know the test has seg fault.
When the signal is caught the handler is fired off to print the message and then
the same signal is sent back again to the process so it can be terminated.
The signal hanlder gets registered before the main() function gets executed
making use of GCC constructor attribute.
The functionality can be enabled defining CTEST_SIG before including ctest.h